Homework sprint 16, correction of LibraryViewModel
