+++ app/build.gradle.kts
@@ -51,12 +51,12 @@ dependencies {
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    implementation("com.google.android.material:material:1.6.1")
-    implementation("com.github.bumptech.glide:glide:4.14.2")
-    annotationProcessor("com.github.bumptech.glide:compiler:4.14.2")
-    implementation("com.squareup.retrofit2:retrofit:${libs.versions.retrofit.get()}")
-    implementation ("com.squareup.retrofit2:converter-gson:${libs.versions.retrofit.get()}")
-    implementation ("androidx.lifecycle:lifecycle-livedata-ktx:2.5.1")
-    implementation ("androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1")
-    implementation ("androidx.activity:activity-ktx:1.6.1")
+    implementation(libs.material)
+    implementation(libs.github.glide)
+    annotationProcessor(libs.compiler)
+    implementation(libs.retrofit2.retrofit)
+    implementation (libs.converter.gson)
+    implementation (libs.androidx.lifecycle.livedata.ktx)
+    implementation (libs.androidx.lifecycle.viewmodel.ktx)
+    implementation (libs.androidx.activity.ktx)
 }
